requirements.txt
discord.py==1.7.3
youtube_dl==2021.6.6
python-dotenv==0.19.0
urllib3==1.26.7
youtube-search-python==1.4.9